wallet-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.215.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.0
pytest-asyncio==0.21.1
requests==2.31.0
python-jose[cryptography]==3.3.0

